refactor(backend): extract media route access guards
